+++ __tests__/0002-smtp-ports.test.ts
@@ -1,4 +1,4 @@
-// SMTP Port Configuration Tests
+// 0002: SMTP Port Configuration Tests
 //
 // Tests different port configurations and behaviors
