+++ __tests__/0003-smtp-sequences.test.ts
@@ -1,11 +1,12 @@
-// SMTP TLS Configuration Tests
+// 0003: SMTP Custom Sequence Tests
 //
-// Tests TLS/SSL configurations and security settings
+// Tests custom SMTP step sequences and protocol control
 
 import { clearDefaultCache } from '../src/cache';
 import { verifyMailboxSMTP } from '../src/smtp';
 import type { SmtpVerificationResult } from '../src/types';
-import { createTestParams, TEST_CONFIGS, TestUtils } from './smtp.test.config';
+import { SMTPStep } from '../src/types';
+import { createTestParams, TEST_SEQUENCES, TestUtils } from './smtp.test.config';
 
 // Helper to map SmtpVerificationResult to boolean|null for legacy assertions
 function toBooleanResult(result: SmtpVerificationResult): boolean | null {
@@ -15,49 +16,126 @@ function toBooleanResult(result: SmtpVerificationResult): boolean | null {
   return result.isDeliverable;
 }
 
-describe('SMTP TLS Configuration', () => {
+describe('SMTP Custom Sequences', () => {
   beforeEach(() => {
     clearDefaultCache();
   });
-  describe('TLS Enable/Disable', () => {
+  describe('Predefined Sequences', () => {
     it(
-      'should work with TLS enabled (default)',
+      'should execute minimal sequence',
       async () => {
         const params = createTestParams({
           options: {
-            ports: [587, 465],
-            tls: true,
+            sequence: TEST_SEQUENCES.MINIMAL,
+            ports: [587],
           },
         });
 
         const smtpResult = await verifyMailboxSMTP(params);
         expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-        console.log(`TLS enabled result: ${toBooleanResult(smtpResult)}`);
+        console.log(`Minimal sequence result: ${toBooleanResult(smtpResult)}`);
       },
       TestUtils.getTestTimeout('integration')
     );
 
     it(
-      'should work with TLS disabled',
-      async () => {
-        const params = createTestParams({
-          options: TEST_CONFIGS.TLS_DISABLED,
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-        console.log(`TLS disabled result: ${toBooleanResult(smtpResult)}`);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with TLS on port 25 (STARTTLS)',
+      'should execute default sequence',
       async () => {
         const params = createTestParams({
           options: {
+            sequence: TEST_SEQUENCES.DEFAULT,
             ports: [25],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+        console.log(`Default sequence result: ${toBooleanResult(smtpResult)}`);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should execute sequence with STARTTLS',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: TEST_SEQUENCES.WITH_STARTTLS,
+            ports: [587],
             tls: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+        console.log(`STARTTLS sequence result: ${toBooleanResult(smtpResult)}`);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should execute sequence with VRFY fallback',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: TEST_SEQUENCES.WITH_VRFY,
+            ports: [25],
+            useVRFY: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+        console.log(`VRFY sequence result: ${toBooleanResult(smtpResult)}`);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should execute full sequence',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: TEST_SEQUENCES.FULL,
+            ports: [587],
+            tls: true,
+            useVRFY: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+        console.log(`Full sequence result: ${toBooleanResult(smtpResult)}`);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+  });
+
+  describe('Custom Sequences', () => {
+    it(
+      'should execute EHLO-only sequence',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: TEST_SEQUENCES.EHLO_ONLY,
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        // EHLO-only should complete successfully but not validate email
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should execute sequence without greeting',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: TEST_SEQUENCES.NO_GREETING,
+            ports: [587],
           },
         });
 
@@ -68,10 +146,219 @@ describe('SMTP TLS Configuration', () => {
     );
 
     it(
-      'should work with TLS on port 587 (STARTTLS)',
+      'should execute VRFY-only test',
       async () => {
         const params = createTestParams({
           options: {
+            sequence: TEST_SEQUENCES.VRFY_ONLY,
+            ports: [25], // VRFY more likely on port 25
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+  });
+
+  describe('Sequence Step Control', () => {
+    it(
+      'should handle single step sequences',
+      async () => {
+        const singleStepSequences = [
+          { steps: [SMTPStep.GREETING], name: 'greeting-only' },
+          { steps: [SMTPStep.EHLO], name: 'ehlo-only' },
+          { steps: [SMTPStep.QUIT], name: 'quit-only' },
+        ];
+
+        for (const seq of singleStepSequences) {
+          const params = createTestParams({
+            options: {
+              sequence: {
+                steps: seq.steps,
+              },
+              ports: [587],
+              timeout: 2000,
+            },
+          });
+
+          const smtpResult = await verifyMailboxSMTP(params);
+          expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+          console.log(`${seq.name} result: ${toBooleanResult(smtpResult)}`);
+        }
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should handle repeated steps',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.EHLO, SMTPStep.EHLO, SMTPStep.QUIT],
+            },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should handle empty sequence gracefully',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [],
+            },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(toBooleanResult(smtpResult)).toBe(true); // Should complete sequence
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+  });
+
+  describe('Custom MAIL FROM Configuration', () => {
+    it(
+      'should use null sender by default',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+              from: '<>',
+            },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should use custom sender email',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+              from: '<sender@example.com>',
+            },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should use sender without brackets',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+              from: 'sender@example.com',
+            },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+  });
+
+  describe('Custom VRFY Configuration', () => {
+    it(
+      'should use local part as VRFY target',
+      async () => {
+        const params = createTestParams({
+          local: 'testuser',
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO, SMTPStep.VRFY],
+              vrfyTarget: 'testuser',
+            },
+            ports: [25],
+            useVRFY: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should use full email as VRFY target',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO, SMTPStep.VRFY],
+              vrfyTarget: 'test@gmail.com',
+            },
+            ports: [25],
+            useVRFY: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should automatically use local part if VRFY target not specified',
+      async () => {
+        const params = createTestParams({
+          local: 'username',
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO, SMTPStep.VRFY],
+            },
+            ports: [25],
+            useVRFY: true,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+  });
+
+  describe('STARTTLS in Sequences', () => {
+    it(
+      'should handle STARTTLS when included in sequence',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.STARTTLS, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+            },
             ports: [587],
             tls: true,
           },
@@ -84,10 +371,32 @@ describe('SMTP TLS Configuration', () => {
     );
 
     it(
-      'should work with TLS on port 465 (implicit)',
+      'should ignore STARTTLS when TLS is disabled',
       async () => {
         const params = createTestParams({
           options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.STARTTLS, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+            },
+            ports: [587],
+            tls: false,
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should handle STARTTLS on port 465',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.STARTTLS, SMTPStep.MAIL_FROM, SMTPStep.RCPT_TO],
+            },
             ports: [465],
             tls: true,
           },
@@ -100,262 +409,23 @@ describe('SMTP TLS Configuration', () => {
     );
   });
 
-  describe('TLS Certificate Validation', () => {
+  describe('Sequence Performance', () => {
     it(
-      'should work with certificate validation disabled (default)',
+      'should compare performance of different sequences',
       async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: {
-              rejectUnauthorized: false,
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with certificate validation enabled',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [465], // More likely to have valid certs
-            tls: {
-              rejectUnauthorized: true,
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should handle certificate validation errors gracefully',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: {
-              rejectUnauthorized: true,
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        // May fail due to cert validation, but should handle gracefully
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-  });
-
-  describe('TLS Version Configuration', () => {
-    it(
-      'should work with TLS 1.2 minimum',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: {
-              minVersion: 'TLSv1.2',
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with TLS 1.3 minimum',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [465], // More likely to support TLS 1.3
-            tls: {
-              minVersion: 'TLSv1.3',
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should handle unsupported TLS version gracefully',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: {
-              minVersion: 'TLSv1.2', // Most servers support this
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-  });
-
-  describe('TLS Configuration Combinations', () => {
-    it(
-      'should work with strict TLS settings',
-      async () => {
-        const params = createTestParams({
-          options: TEST_CONFIGS.TLS_STRICT,
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-        console.log(`Strict TLS result: ${toBooleanResult(smtpResult)}`);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with lenient TLS settings',
-      async () => {
-        const params = createTestParams({
-          options: TEST_CONFIGS.TLS_LENIENT,
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-        console.log(`Lenient TLS result: ${toBooleanResult(smtpResult)}`);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with custom TLS configuration',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: {
-              rejectUnauthorized: false,
-              minVersion: 'TLSv1.2',
-            },
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-  });
-
-  describe('STARTTLS Behavior', () => {
-    it(
-      'should upgrade to TLS on port 25 when STARTTLS is available',
-      async () => {
-        const consoleSpy = jest.spyOn(console, 'log').mockImplementation();
-
-        const params = createTestParams({
-          options: {
-            ports: [25],
-            tls: true,
-            debug: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-
-        // Check for TLS upgrade logs
-        const logs = consoleSpy.mock.calls.flat().join(' ');
-        consoleSpy.mockRestore();
-
-        // TLS upgrade might be mentioned in logs
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should upgrade to TLS on port 587',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should not attempt STARTTLS on port 465 (implicit TLS)',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [465],
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should continue without TLS if STARTTLS fails',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [25],
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-  });
-
-  describe('TLS Performance', () => {
-    it(
-      'should compare TLS vs non-TLS performance',
-      async () => {
-        const testCases = [
-          { name: 'No TLS', tls: false, port: 25 },
-          { name: 'TLS Port 25', tls: true, port: 25 },
-          { name: 'TLS Port 587', tls: true, port: 587 },
-          { name: 'TLS Port 465', tls: true, port: 465 },
+        const sequences = [
+          { name: 'Minimal', seq: TEST_SEQUENCES.MINIMAL },
+          { name: 'Default', seq: TEST_SEQUENCES.DEFAULT },
+          { name: 'With STARTTLS', seq: TEST_SEQUENCES.WITH_STARTTLS },
         ];
 
         const results: { [key: string]: { result: boolean | null; duration: number } } = {};
 
-        for (const testCase of testCases) {
+        for (const { name, seq } of sequences) {
           const params = createTestParams({
             options: {
-              ports: [testCase.port],
-              tls: testCase.tls,
+              sequence: seq,
+              ports: [587],
               debug: false,
             },
           });
@@ -364,100 +434,30 @@ describe('SMTP TLS Configuration', () => {
           const smtpResult = await verifyMailboxSMTP(params);
           const duration = Date.now() - start;
 
-          results[testCase.name] = { result: toBooleanResult(smtpResult), duration };
-          console.log(`${testCase.name}: ${duration}ms`);
+          results[name] = { result: toBooleanResult(smtpResult), duration };
+          console.log(`${name} sequence: ${duration}ms, result: ${toBooleanResult(smtpResult)}`);
         }
 
-        // Verify all results are valid
+        // Verify all sequences complete successfully
         Object.values(results).forEach(({ result }) => {
           expect(TestUtils.isValidResult(result)).toBe(true);
         });
       },
       TestUtils.getTestTimeout('slow')
     );
-
-    it(
-      'should handle TLS handshake timeout',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [465],
-            tls: {
-              rejectUnauthorized: true,
-            },
-            timeout: 1000, // Short timeout for TLS handshake
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
   });
 
-  describe('TLS Security Scenarios', () => {
+  describe('Error Handling in Sequences', () => {
     it(
-      'should prefer secure ports when TLS is enabled',
+      'should handle invalid step in sequence',
       async () => {
         const params = createTestParams({
           options: {
-            ports: [465, 587, 25], // Secure ports first
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should work with mixed secure/unsecure ports',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [25, 587, 465], // Mixed order
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-
-    it(
-      'should handle TLS on servers that dont support it',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [25],
-            tls: true,
-          },
-        });
-
-        const smtpResult = await verifyMailboxSMTP(params);
-        // Should either succeed with TLS or fall back to plain text
-        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
-      },
-      TestUtils.getTestTimeout('integration')
-    );
-  });
-
-  describe('TLS with Custom Sequences', () => {
-    it(
-      'should handle STARTTLS in custom sequence',
-      async () => {
-        const params = createTestParams({
-          options: {
-            ports: [587],
-            tls: true,
             sequence: {
-              steps: ['GREETING', 'EHLO', 'STARTTLS', 'MAIL_FROM', 'RCPT_TO'] as any,
+              steps: [SMTPStep.QUIT, SMTPStep.EHLO], // Invalid order
             },
+            ports: [587],
+            timeout: 2000,
           },
         });
 
@@ -468,15 +468,33 @@ describe('SMTP TLS Configuration', () => {
     );
 
     it(
-      'should not use STARTTLS if not in sequence',
+      'should handle sequence without MAIL_FROM',
       async () => {
         const params = createTestParams({
           options: {
-            ports: [587],
-            tls: true,
             sequence: {
-              steps: ['GREETING', 'EHLO', 'MAIL_FROM', 'RCPT_TO'] as any, // No STARTTLS
+              steps: [SMTPStep.GREETING, SMTPStep.EHLO, SMTPStep.RCPT_TO], // Missing MAIL_FROM
             },
+            ports: [587],
+          },
+        });
+
+        const smtpResult = await verifyMailboxSMTP(params);
+        expect(TestUtils.isValidResult(toBooleanResult(smtpResult))).toBe(true);
+      },
+      TestUtils.getTestTimeout('integration')
+    );
+
+    it(
+      'should handle sequence with only QUIT',
+      async () => {
+        const params = createTestParams({
+          options: {
+            sequence: {
+              steps: [SMTPStep.QUIT],
+            },
+            ports: [587],
+            timeout: 1000,
           },
         });
