+++ __tests__/0007-is-email-exists-core.test.ts
@@ -1,3 +1,7 @@
+// 0007: Core Email Verification Tests
+//
+// Tests the core is-email-exists functionality ported from Rust
+
 import dns from 'node:dns';
 import {
   EmailProvider,
